Rename s3_ls -> s3
